+++ src/uart/uart.h
@@ -9,7 +9,7 @@
 #define UART_RX_STREAM_SIZE  512U
 
 void uart_init(void);
-void uart_transmit(const uint8_t *buf, uint16_t len);
+void uart_transmit(const uint8_t *buf, size_t len);
 uint32_t uart_get_drop_count(void);
 
 /* Protocol layer reads raw bytes from here. Blocks until data is available. */
